Move projects app queries to its own folder
